Move relengapi.celery to relengapi.lib.celery
..and clean up some docs.  Part of #86.
