+++ base/docs/deployment/@relengapi/workers.rst
@@ -1,5 +1,5 @@
-Starting Workers
-================
+Starting Celery Workers
+=======================
 
 Releng API uses `Celery <http://www.celeryproject.org/>`_ to distribute tasks to workers.
 Releng API workers are simply Celery workers invoked with ::
